requirements.txt
datasets==3.3.2
fastapi==0.115.6
pypandoc_binary==1.15
python-dotenv==1.0.1
requests==2.32.3
sentry-sdk==2.22.0
streamlit==1.43.2
# torch==2.6.0
transformers==4.49.0
uvicorn==0.34.0
